refactored appointment list item component
component accepts child components
