+++ src/applications/vaos/appointment-list/components/PastAppointmentsListV2/index.jsx
@@ -3,13 +3,19 @@ import { shallowEqual, useDispatch, useSelector } from 'react-redux';
 
 import recordEvent from 'platform/monitoring/record-event';
 import moment from 'moment';
+import { focusElement } from 'platform/utilities/ui';
+import { useHistory } from 'react-router-dom';
 import { getPastAppointmentListInfo } from '../../redux/selectors';
 import {
   FETCH_STATUS,
   GA_PREFIX,
   APPOINTMENT_TYPES,
+  SPACE_BAR,
 } from '../../../utils/constants';
-import { getVAAppointmentLocationId } from '../../../services/appointment';
+import {
+  getLink,
+  getVAAppointmentLocationId,
+} from '../../../services/appointment';
 import AppointmentListItem from '../AppointmentsPageV2/AppointmentListItem';
 import NoAppointments from '../NoAppointments';
 import PastAppointmentsDateDropdown from './PastAppointmentsDateDropdown';
@@ -19,8 +25,30 @@ import {
   fetchPastAppointments,
   startNewAppointmentFlow,
 } from '../../redux/actions';
-import { selectFeatureAppointmentList } from '../../../redux/selectors';
+import {
+  selectFeatureAppointmentList,
+  selectFeatureStatusImprovement,
+} from '../../../redux/selectors';
 import AppointmentListGroup from '../AppointmentsPageV2/AppointmentListGroup';
+import Card from '../AppointmentsPageV2/Card';
+
+function handleClick({ history, link, idClickable }) {
+  return () => {
+    if (!window.getSelection().toString()) {
+      focusElement(`#${idClickable}`);
+      history.push(link);
+    }
+  };
+}
+
+function handleKeyDown({ history, link, idClickable }) {
+  return event => {
+    if (!window.getSelection().toString() && event.keyCode === SPACE_BAR) {
+      focusElement(`#${idClickable}`);
+      history.push(link);
+    }
+  };
+}
 
 export function getPastAppointmentDateRangeOptions(today = moment()) {
   const startOfToday = today.clone().startOf('day');
@@ -91,6 +119,7 @@ export function getPastAppointmentDateRangeOptions(today = moment()) {
 }
 
 export default function PastAppointmentsListNew() {
+  const history = useHistory();
   const dispatch = useDispatch();
   const [isInitialMount, setInitialMount] = useState(true);
   const dateRangeOptions = getPastAppointmentDateRangeOptions();
@@ -104,6 +133,9 @@ export default function PastAppointmentsListNew() {
   } = useSelector(state => getPastAppointmentListInfo(state), shallowEqual);
   const featureAppointmentList = useSelector(state =>
     selectFeatureAppointmentList(state),
+  );
+  const featureStatusImprovement = useSelector(state =>
+    selectFeatureStatusImprovement(state),
   );
 
   useEffect(() => {
@@ -235,6 +267,11 @@ export default function PastAppointmentsListNew() {
               >
                 {monthBucket.map((appt, index) => {
                   const facilityId = getVAAppointmentLocationId(appt);
+                  const idClickable = `id-${appt.id.replace('.', '\\.')}`;
+                  const link = getLink({
+                    featureStatusImprovement,
+                    appointment: appt,
+                  });
 
                   if (
                     appt.vaos.appointmentType ===
@@ -246,8 +283,20 @@ export default function PastAppointmentsListNew() {
                       <AppointmentListItem
                         key={index}
                         appointment={appt}
-                        facility={facilityData[facilityId]}
-                      />
+                        className="vaos-appts__card--clickable vads-u-margin-bottom--3"
+                      >
+                        <Card
+                          appointment={appt}
+                          facility={facilityData[facilityId]}
+                          link={link}
+                          handleClick={() =>
+                            handleClick({ history, link, idClickable })
+                          }
+                          handleKeyDown={() =>
+                            handleKeyDown({ history, link, idClickable })
+                          }
+                        />
+                      </AppointmentListItem>
                     );
                   }
                   return null;
